+++ cbits/proteaaudio_binding.h
@@ -23,7 +23,7 @@ sound_t soundLoop(sample_t sample, float volumeL, float volumeR, float disparity
 sound_t soundPlay(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
 sound_t soundPlayOn(unsigned int track, sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
 sound_t soundLoopOn(unsigned int track, sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
-int soundUpdate(sound_t sound, float volumeL, float volumeR, float disparity, float pitch);
+int soundUpdate(sound_t sound, int pause, float volumeL, float volumeR, float disparity, float pitch);
 int soundStop(sound_t sound);
 int soundActive(sound_t sound);
 #ifdef __cplusplus