+++ proteaaudio/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.0
 - add MP3 support
+- add pause support: soundUpdate takes a Bool (is paused) parameter
 
 # 0.9.4
 - fix PulseAudio default output device detection